Feature/internationalization (#8)
* feat: Add support for internationalization (i18n)

The code changes introduce support for internationalization (i18n) in the application. The following updates are made:
- Added new files: `src/locales/en.json` and `src/locales/es.json` to store translation resources for English and Spanish languages respectively.
- Created `src/types/types.d.ts` and `src/types/resources.d.ts` to define types for translation resources.
- Added `i18n.ts` file to initialize i18n with translation resources and language options.
- Updated `Drawer.tsx` component to use translations for menu items.
- Added `useTranslation` hook from `react-i18next` to translate text in components.

These changes enable the application to support multiple languages and improve the user experience for international users.

* feat: Add translations for dashboard page

This commit adds translations for the dashboard page in both English and Spanish. The `en.json` and `es.json` files in the `src/locales` directory are updated with the translated text for the dashboard section. The translations include the titles for various analytics cards such as "Total Savings," "Total Fat Burn," "Total Daily Goals Streak," and "Total P/L Trading." This enhancement improves the user experience by providing localized content for international users.

* feat: Add translations for personal finances page

This commit adds translations for the personal finances page in both English and Spanish. The `en.json` and `es.json` files in the `src/locales` directory are updated with the translated text for the personal finances section. The translations include the title for the personal finances section as well as the labels for the savings and debts graph. This enhancement improves the user experience by providing localized content for international users.