Feature/finances-personal (#12)
* feat: Add PageHeader component to common directory

The code changes introduce the following updates:
- Added the PageHeader component to the common directory.
- The PageHeader component is a reusable component that wraps its children in a Box component with a margin-bottom of 8.
- The PageHeader component is exported from the PageHeader.tsx file.

These changes add a new component, PageHeader, to the common directory, which can be used to display a header section on pages throughout the application.

* feat: Add number, auth, and date utility functions

The code changes introduce the following updates:
- Added number utility functions for formatting currency and percentages.
- Added auth utility functions for handling user authentication.
- Added date utility functions for formatting dates.
- Exported the number, auth, and date utility functions from the utils/index.ts file.

These changes add utility functions for common operations related to numbers, authentication, and dates, improving code reusability and maintainability.

* refactor: Update DebtCard component to use useDebts hook for data management

* refactor: Remove unused exports from finances-related files

* refactor: Remove unused exports and update TradingJournalPage

The code changes remove the unused export from the utils/date.ts file. Additionally, the TradingJournalPage component is updated to use the translated title for the finances section. These changes improve code cleanliness and ensure consistency in the application's user interface.

* refactor: Update minHeight in PageHeader, FixedExpenseCard, DebtCard, and IncomeCard components

The code changes update the minHeight property in the PageHeader, FixedExpenseCard, DebtCard, and IncomeCard components. The minHeight value is increased from 170 to 175 in order to provide a slightly larger height for these components. This change ensures that the components have enough space to display their content properly, improving the overall user experience.

* refactor: Update personal finance related components and translations

The code changes update the personal finance related components and translations. The PersonalFinances.hooks.ts file is updated to use the IPersonalFinance interface instead of IPersonalFinances. The addPersonalFinance function is added to the usePersonalFinances hook to handle adding new personal finance data. The en.json and resources.d.ts files are updated with new translations for the "addPlan" and "addPlanHint" strings. These changes improve the code structure and provide better support for managing personal finance data and translations.

* refactor: Add date utility function and update financial models

* ```text
refactor: Update PersonalFinancesPage to create new plan with incremented name

The code changes update the PersonalFinancesPage component to create a new plan with an incremented name when the "Create New Plan" button is clicked. This improves the functionality of the page by ensuring that each new plan has a unique name.

* refactor: Update PersonalFinancesPage to create new plan with incremented name

* refactor: Disable require-jsdoc and update import/no-unresolved rule in .eslintrc.js

This commit disables the require-jsdoc rule and updates the import/no-unresolved rule in the .eslintrc.js file. The require-jsdoc rule is turned off to allow for more flexibility in documenting code with JSDoc comments. The import/no-unresolved rule is set to warn to catch any unresolved imports and prevent potential runtime errors. These changes improve the linting configuration for the project.

* refactor: Update DebtCard component to conditionally render DebtEditDialog

The code changes update the DebtCard component to conditionally render the DebtEditDialog based on the value of the "canEdit" prop. If "canEdit" is true, the DebtEditDialog is rendered, allowing the user to edit the debt data. If "canEdit" is false, the DebtEditDialog is not rendered, preventing the user from editing the debt data. This change improves the flexibility and usability of the DebtCard component.

* ```text
refactor: Update PersonalFinancesPage to handle multiple financial plans

The code changes update the PersonalFinancesPage component to handle multiple financial plans. The _updateDebts function is modified to update the debts for all financial plans if a planId is not provided. This change improves the functionality of the page by allowing users to update debts for all plans at once.

* refactor: Update PersonalFinancesPage to handle multiple financial plans

* refactor: expenses and incomes out of financialsnapshot

* feat: animation for tabs

* refactor: Improve PersonalFinancesPage UX with updated styles and hover effect

The code changes update the PersonalFinancesPage component to improve the user experience (UX). The styles for the back button and page title are updated to use the "info.main" color and have a hover effect that underlines the text. This change enhances the visual appeal and interactivity of the page.

* refactor: Update PersonalFinancesPage to handle multiple financial plans

The code changes update the PersonalFinancesPage component to handle multiple financial plans. The _updateDebts function is modified to update the debts for all financial plans if a planId is not provided. This change improves the functionality of the page by allowing users to update debts for all plans at once.

* refactor: Update financial plan types and components to include name fields

The code changes update the financial plan types and components to include name fields for debts, incomes, and fixed expenses. This change allows users to provide names for their financial items, improving the clarity and organization of the data.

* refactor: Sort incomes in IncomeEditDialog and PersonalFinancesGraph

The code changes sort the incomes in the IncomeEditDialog and PersonalFinancesGraph components. In the IncomeEditDialog component, the incomes are sorted in descending order based on the amount. In the PersonalFinancesGraph component, the incomes are sorted in descending order based on the singleDate. This change improves the consistency and readability of the displayed data.